Move sign in to pages folder
